backend/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
yfinance==0.2.28
pandas==2.0.3
numpy==2.1.1
google-generativeai==0.3.2
textblob==0.17.1
requests==2.31.0
aiohttp==3.8.5
python-dotenv==1.0.0
